SDS011: change type of update_interval_min to uint8_t

+++ src/esphome/sensor/sds011_component.h
@@ -23,7 +23,7 @@ class SDS011Component : public Component, public UARTDevice {
    * @param update_interval_min The update interval in minutes.
    * @param rx_mode_only RX-only mode to avoid sending data to the sensor.
    */
-  SDS011Component(UARTComponent *parent, uint32_t update_interval_min = 0, bool rx_mode_only = false);
+  SDS011Component(UARTComponent *parent, uint8_t update_interval_min = 0, bool rx_mode_only = false);
 
   /// Manually set the rx-only mode. Defaults to false.
   void set_rx_mode_only(bool rx_mode_only);
@@ -39,7 +39,7 @@ class SDS011Component : public Component, public UARTDevice {
   SDS011Sensor *make_pm_10_0_sensor(const std::string &name);
   bool get_rx_mode_only() const;
 
-  void set_update_interval_min(uint32_t update_interval_min);
+  void set_update_interval_min(uint8_t update_interval_min);
 
  protected:
   void sds011_write_command_(const uint8_t *command);
@@ -54,7 +54,7 @@ class SDS011Component : public Component, public UARTDevice {
   uint8_t data_[10];
   uint8_t data_index_{0};
   uint32_t last_transmission_{0};
-  uint32_t update_interval_min_;
+  uint8_t update_interval_min_;
 
   bool rx_mode_only_;
 };
